# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/book": {
            "post": {
                "description": "create book",
                "tags": [
                    "book"
                ],
                "summary": "create book",
                "parameters": [
                    {
                        "description": "body",
                        "name": "book",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_yanarowana123_onelab2_internal_models.CreateBookRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_yanarowana123_onelab2_internal_models.BookResponse"
                        }
                    }
                }
            }
        },
        "/book/{bookID}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "get book by id",
                "tags": [
                    "book"
                ],
                "summary": "get book by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Book ID (UUID format)",
                        "name": "bookID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_yanarowana123_onelab2_internal_models.BookResponse"
                        }
                    },
                    "404": {
                        "description": "book not found"
                    }
                }
            }
        },
        "/checkout/{bookID}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "checkout book",
                "tags": [
                    "checkOut"
                ],
                "summary": "checkout book",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Book ID (UUID format)",
                        "name": "bookID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "you already have checked out this book"
                    },
                    "404": {
                        "description": "book not found"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "log-in to system",
                "tags": [
                    "auth"
                ],
                "summary": "log-in to system",
                "parameters": [
                    {
                        "description": "body",
                        "name": "auth",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_yanarowana123_onelab2_internal_models.AuthUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_yanarowana123_onelab2_internal_models.Tokens"
                        }
                    }
                }
            }
        },
        "/return/{bookID}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "return book",
                "tags": [
                    "checkOut"
                ],
                "summary": "return book",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Book ID (UUID format)",
                        "name": "bookID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "book not found"
                    }
                }
            }
        },
        "/signup": {
            "post": {
                "description": "Registration",
                "tags": [
                    "auth"
                ],
                "summary": "Registration",
                "parameters": [
                    {
                        "description": "body",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_yanarowana123_onelab2_internal_models.CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_yanarowana123_onelab2_internal_models.UserResponse"
                        }
                    },
                    "400": {
                        "description": "validation error"
                    }
                }
            }
        },
        "/user/{userID}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "get user by id",
                "tags": [
                    "user"
                ],
                "summary": "get user by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID (UUID format)",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_yanarowana123_onelab2_internal_models.UserResponse"
                        }
                    },
                    "404": {
                        "description": "user not found"
                    }
                }
            }
        },
        "/users/book-quantity": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "get user list with books quantity",
                "tags": [
                    "user"
                ],
                "summary": "get user list with books quantity they have at the moment",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page (pagination)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page size (pagination)",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_yanarowana123_onelab2_internal_models.UserWithBookQuantityList"
                        }
                    }
                }
            }
        },
        "/users/books": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "get user list with books",
                "tags": [
                    "user"
                ],
                "summary": "get user list with books",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page (pagination)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page size (pagination)",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_yanarowana123_onelab2_internal_models.UserWithBookList"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_yanarowana123_onelab2_internal_models.AuthUser": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "github_com_yanarowana123_onelab2_internal_models.BookResponse": {
            "type": "object",
            "properties": {
                "author_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "github_com_yanarowana123_onelab2_internal_models.CreateBookRequest": {
            "type": "object",
            "required": [
                "author_id",
                "name"
            ],
            "properties": {
                "author_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "github_com_yanarowana123_onelab2_internal_models.CreateUserRequest": {
            "type": "object",
            "required": [
                "email",
                "first_name",
                "last_name",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "github_com_yanarowana123_onelab2_internal_models.Tokens": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "github_com_yanarowana123_onelab2_internal_models.UserResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                }
            }
        },
        "github_com_yanarowana123_onelab2_internal_models.UserWithBook": {
            "type": "object",
            "properties": {
                "books": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_yanarowana123_onelab2_internal_models.BookResponse"
                    }
                },
                "user": {
                    "$ref": "#/definitions/github_com_yanarowana123_onelab2_internal_models.UserResponse"
                }
            }
        },
        "github_com_yanarowana123_onelab2_internal_models.UserWithBookList": {
            "type": "object",
            "properties": {
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_yanarowana123_onelab2_internal_models.UserWithBook"
                    }
                }
            }
        },
        "github_com_yanarowana123_onelab2_internal_models.UserWithBookQuantity": {
            "type": "object",
            "properties": {
                "books": {
                    "type": "integer"
                },
                "user": {
                    "$ref": "#/definitions/github_com_yanarowana123_onelab2_internal_models.UserResponse"
                }
            }
        },
        "github_com_yanarowana123_onelab2_internal_models.UserWithBookQuantityList": {
            "type": "object",
            "properties": {
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_yanarowana123_onelab2_internal_models.UserWithBookQuantity"
                    }
                }
            }
        }
    }
}